+++ source/renderer/app/components/widgets/AssetToken.js
@@ -7,7 +7,7 @@ import { defineMessages, intlShape } from 'react-intl';
 import CopyToClipboard from 'react-copy-to-clipboard';
 import { observer } from 'mobx-react';
 import styles from './AssetToken.scss';
-import { ellipsis } from '../../utils/strings';
+import { ellipsis, hexToString } from '../../utils/strings';
 import type { WalletSummaryAsset } from '../../api/assets/types';
 import copyIcon from '../../assets/images/copy-asset.inline.svg';
 import copyCheckmarkIcon from '../../assets/images/check-w.inline.svg';
@@ -159,12 +159,7 @@ export default class AssetToken extends Component<Props, State> {
     );
   }
 
-  assetItemRenderer = (
-    assetId: string,
-    assetItem: string,
-    value: string,
-    singleline?: boolean
-  ) => {
+  assetItemRenderer = (assetId: string, assetItem: string, value: string) => {
     const { itemCopied } = this.state;
     const icon = itemCopied === assetId ? copyCheckmarkIcon : copyIcon;
     const iconClassnames = classnames([
@@ -177,10 +172,15 @@ export default class AssetToken extends Component<Props, State> {
     return (
       <CopyToClipboard text={value} onCopy={onCopy}>
         <div className={styles.assetItem}>
-          <em className={singleline ? styles.singleline : null}>
+          <div className={styles.value}>
             {value}
             <SVGInline svg={icon} className={iconClassnames} />
-          </em>
+          </div>
+          {assetId === 'assetName' && (
+            <div className={styles.assetASCIIName}>
+              (ASCII: {hexToString(value)})
+            </div>
+          )}
         </div>
       </CopyToClipboard>
     );
@@ -190,7 +190,7 @@ export default class AssetToken extends Component<Props, State> {
     const { intl } = this.context;
     const { asset } = this.props;
     const { fingerprint, policyId, assetName, metadata } = asset;
-    const { name, acronym, description } = metadata || {};
+    const { name, ticker, description } = metadata || {};
     const item = this.assetItemRenderer;
     return (
       <div className={styles.tooltipContent}>
@@ -202,14 +202,14 @@ export default class AssetToken extends Component<Props, State> {
           )}
         </div>
         <dl>
-          {acronym && (
+          {ticker && (
             <F>
               <dt>{intl.formatMessage(messages.tickerItem)}</dt>
               <dd>
                 {item(
-                  'acronym',
+                  'ticker',
                   intl.formatMessage(messages.tickerItem),
-                  acronym
+                  ticker
                 )}
               </dd>
             </F>
@@ -288,11 +288,11 @@ export default class AssetToken extends Component<Props, State> {
           }}
           contentClassName={styles.popOver}
           content={tooltipContent}
-          isShowingOnHover={false}
           visible={isTooltipVisible}
           appendTo="parent"
           maxWidth={376}
           allowHTML
+          interactive
         >
           {children}
         </PopOver>