+++ source/renderer/app/utils/shelleyTrezor.js
@@ -26,7 +26,7 @@ export const prepareTrezorOutput = (output: CoinSelectionOutput) => {
     // Change output
     return {
       amount: output.amount.quantity.toString(),
-      tokenBundle: _getAssets(output.assets),
+      tokenBundle: _getTokenBundle(output.assets),
       addressParameters: {
         addressType: 0, // BASE address
         path: derivationPathToString(output.derivationPath),
@@ -37,7 +37,7 @@ export const prepareTrezorOutput = (output: CoinSelectionOutput) => {
   return {
     address: output.address,
     amount: output.amount.quantity.toString(),
-    tokenBundle: _getAssets(output.assets),
+    tokenBundle: _getTokenBundle(output.assets),
   };
 };
 
@@ -66,14 +66,16 @@ export const prepareTrezorWithdrawal = (
 
 // Helper Methods
 
-const _getAssets = (assets) => {
+const _getTokenBundle = (assets) => {
   const constructedAssets = map(assets, (asset) => {
     return {
       policyId: asset.policyId,
-      tokens: {
-        assetNameBytes: asset.assetName,
-        amount: asset.quantity.toString(),
-      },
+      tokenAmounts: [
+        {
+          assetNameBytes: asset.assetName,
+          amount: asset.quantity.toString(),
+        },
+      ],
     };
   });
   return constructedAssets;